requirements.txt
langgraph>=0.2.0
langchain-groq>=0.1.9
langchain-openai>=0.1.0
langchain-core>=0.2.0
gradio>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
